refactor: split Chart component into CandlestickChart to make it reusable (Mobile)
