+++ src/components/mobile/Chart.mobile.vue
@@ -3,7 +3,7 @@
     <ChartOptions :crypto="crypto" :base="base" v-model="base" />
     <div class="stats">
       <span class="price">
-        <span class="h2">$95.23</span>
+        <h2>$95.23</h2>
       </span>
       <span class="statsgroup">
         <span>% Change</span>
@@ -18,68 +18,42 @@
         <span class="red">75.20</span>
       </span>
     </div>
-    <VChart :option="option" style="margin-bottom: 1em" autoresize />
+    <CandlestickChart :crypto="crypto" :base="base" />
     <slot />
   </section>
 </template>
 
 
-<script>
+<script setup>
 import ChartOptions from "@/components/mobile/ChartOptions.mobile.vue";
-import { onMounted, reactive, ref } from "vue";
-import options from "@/config/chartOptions.mobile.js";
+import CandlestickChart from "@/components/mobile/CandlestickChart.mobile.vue";
+import { ref } from "vue";
 
-import { use } from "echarts/core";
-import { CanvasRenderer } from "echarts/renderers";
-import { CandlestickChart } from "echarts/charts";
-import { GridComponent, DataZoomComponent, TooltipComponent } from "echarts/components";
-import VChart from "vue-echarts";
-
-export default {
-  props: {
-    crypto: {
-      type: String,
-      required: true
-    },
-
-    base: {
-      type: String,
-      required: true
-    }
+const props = defineProps({
+  crypto: {
+    type: String,
+    required: true
   },
 
-  components: { VChart, ChartOptions },
-
-  setup(props) {
-    use([CandlestickChart, CanvasRenderer, GridComponent, DataZoomComponent, TooltipComponent ]);
-
-    const { crypto } = props;
-    const base = ref(props.base);
-    const option = reactive(options);
-
-    // Temporary load the data
-    onMounted(() => {
-      fetch("https://api.binance.com/api/v3/klines?symbol=ETHEUR&interval=5m")
-        .then(res => res.json())
-        .then(data => {
-          data = data.slice(-100, -1);
-          // Get first 5 elements and convert to float
-          data = data.map(candle => candle.slice(0, 5).map(c => parseFloat(c)));
-          option.series.data = data;
-      });
-    });
-
-    return { crypto, base, option }
+  base: {
+    type: String,
+    required: true
   }
-}
+});
+
+const { crypto } = props;
+const base = ref(props.base);
 </script>
 
-<style lang="sass">
+<style lang="sass" scoped>
   .chart-container
     position: absolute
     width: 98vw
     left: 1vw
     height: 60vh
+
+  h2
+    font-weight: normal
 
   .stats
     background: $bg-dark
@@ -97,11 +71,7 @@ export default {
     display: inline-flex
     flex-direction: column
     justify-content: center
-    
-  canvas
-    border-radius: 0 0 1.5em 1.5em
-    cursor: crosshair
-
+  
   .green
     color: $green
