+++ src/config/candlestickChartOptions.mobile.js
